requirements 2.txt
alabaster==0.7.12
appnope==0.1.2
asttokens==2.0.5
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
debugpy==1.6.0
decorator==5.1.1
docutils==0.17.1
entrypoints==0.4
executing==0.8.3
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.3
ipykernel==6.10.0
ipython==8.2.0
jedi==0.18.1
Jinja2==3.1.1
jupyter-client==7.2.0
jupyter-core==4.9.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.22.3
packaging==21.3
pandas==1.4.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.9.0
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
pyzmq==22.3.0
requests==2.27.1
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.1
Sphinx==4.5.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.2.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.9
validators==0.18.2
wcwidth==0.2.5
zipp==3.7.0
transformers
torchsphinx-autodoc-typehints
python-dotenv
requests-random-user-agent
